removed shipment model
