--- app/models/shipment.rb
@@ -1,21 +1,0 @@
-class Shipment < ActiveRecord::Base
-  has_many :packages
-  has_many :estimates
-
-  def get_quotes(origin, destination, packages)
-    quotes = []
-
-    ups = ActiveShipping::UPS.new(login: ENV["UPS_LOGIN"], password: ENV["UPS_PASSWORD"], key: ENV["UPS_KEY"])
-    response = ups.find_rates(origin, destination, packages)
-    ups_rates = response.rates.sort_by(&:price).collect {|rate| [rate.service_name, rate.price]}
-
-    usps = ActiveShipping::USPS.new(login: ENV["ACTIVESHIPPING_USPS_LOGIN"])
-    response = usps.find_rates(origin, destination, packages)
-    usps_rates = response.rates.sort_by(&:price).collect {|rate| [rate.service_name, rate.price ]}
-
-    quotes << ups_rates
-    quotes << usps_rates
-    return quotes
-  end
-
-end
