--- db/migrate/20160119231210_create_shipments.rb
@@ -1,8 +1,0 @@
-class CreateShipments < ActiveRecord::Migration
-  def change
-    create_table :shipments do |t|
-
-      t.timestamps null: false
-    end
-  end
-end
